Tests

+++ app/src/main/java/com/sinergia/todov2/Task.java
@@ -14,17 +14,19 @@ public class Task {
     }
 */
     @PrimaryKey(autoGenerate = true)
-    private int id;
+    private long id;
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 
+
     @NonNull
+
     @ColumnInfo(name = "task_title")
     private String title;
